Rename AppWindow -> Window, add BrowserWebView.
With the addition of the BrowserWebView class, it should now be possible
to embed the webview inside GtkBuilder XML UI description files. It did
not seem to be possible to put a WebKitWebView inside a GtkBuilder UI
file. This might have something to do with there not being a webkit2gtk
glade file in /usr/share/glade/.