Add browser tab-notebook.
The notebook is currently hacked into the window, but it should be added
to the window's UI XML file.
In order to have the notebook's web-view created by gtk-builder, it
seems necessary to call g_type_ensure in the widget before
gtk_widget_template_init. See:
http://stackoverflow.com/questions/24235937/custom-gtk-widget-with-template-ui
for more details.
Also, the type macros seem to be formed wrong. They should probably be
changed to BROWSER_TYPE_* from BROWSER_*_TYPE, to match the Gtk style
for type macros.