Add BrowserTab class.
The BrowserTab class is the main content container for each tab. It's
only child at the moment is a BrowserWebView, but it might hold other
children in the future such as the web-inspector, a statusbar at the
bottom, and any overlay items.
The BrowserWindow and BrowserNotebook do not use any hacks to create
children now, which is good too!