Add basic title syncing between tab & label.
The tab-label updates its title whenever the tab's title changes. The
tab listens to its webview's notify::title signal and then propagates
the signal again with its own notifyt::title. This might be changes in
the future to eliminate the intermediate signal step, but this allows
for more control of the tab/tab-label interaction.

In order to enable tab/tab-label title syncing, the tab and tab-label
could not remain the notebook's UI description, since the tab-label
expects to receive a BrowserTab when created. This might be doable in
the GtkBuilder template, but it isn't really necessary since tabs will
be created programmatically in the future (once the add button, open in
new tab, etc. are supported).