Add entry-activated signal to BrowserToolbar.
Similar to the tab-label's close-clicked signal, the toolbar now has an
entry-activated signal. This simply forwards the activate signal from
the entry.
This signal-forwarding structure is useful and desirable so
that the top-level window can connect to the toolbar's signals and not
have to dig around in the toolbar's widget hierarchy. This would not
work anyways since all the widgets are opaque structs. So, to connect to
the widgets inside the toolbar, the toolbar would need to expose getters
for all its children.