Connect window to toolbar entry-activated signal.
This signal is passed up to the window from the toolbar's entry,
allowing the window to tell the web-view to navigate to the new URI.

Something that needs to be handled is not clobbering the user-entered
text when the web-view updates its URI as part of the page loading
process. So, the proposed solution is: keep track in the toolbar whether
the entry has received user input and store this as a flag. The window
will receive the "uri-changed" (or similar) signal from the
web-view/tab, check whether the toolbar entry has been edited by the
user, and then update the URI if there have been no changes. The "entry
modified" flag in the toolbar will be reset when the user activates the
entry (presses enter) or when there is a "hard" navigation in the tab
(reload, click a link, etc.).