Track changes to the toolbar entry.
As discussed previously, there is a need for a flag in the toolbar to
indicate whether the entry has been changed by the user. Currently, the
"entry_modified" flag is set whenever text is inserted or deleted, and
then unset when the entry is activated.
There will be some cases when the entry should be changed externally
even when there are user-changes, like when the page is reloaded or a
navigation action occurs in the web-view. In these cases, the window can
decide whether to force the entry to be altered, which should also unset
the modified flag in the toolbar.