Add basic uri opening and tab creation.
The window can now create tabs in the notebook from URIs passed in from
the app or elsewhere. The notebook does not add a hacky tab in its init
anymore, which is good!
The next steps are to put the current URI in the toolbar's entry, hook
up the new-tab button, and handle opening of "new windows" in a new tab.