Enable loading uri from entry into tab.
The tab cannot receive focus itself, since it is a container widget (it
could, but we don't want it to get focus, instead we want its children
to get it), so the "grab-focus" signal on the tab is used to forward the
grab to the web-view. In the future, if there are multiple child
widgets, perhaps letting the tab get focus will in turn focus the last
used child widget?