Connect notebook "new-tab" signal to window.
The notebook's new-tab button has its click signal forwarded to the
window, which creates a blank tab.

There was an issue with the order and interleaving of signals for tab
insertion and window updating. When tab is inserted into the notebook, a
signal is sent to the window. However, at this time, the current-page is
still the old one, since the notebook's add_tab function is still
halfway done. So, the window should not update its state (URI, title,
etc.) at this point. At the end of the notebook's add_tab function, the
current page is set (if the jump_to flag is also set).

The preferred place to update the window's state following a tab creation
is when the "switch-page" signal is emitted by the notebook. However,
again the current-page of the notebook is not updated yet, so we actually
have to connect_after the "switch-page" signal. Then there doesn't need
to be any extra state stored by the notebook (at the moment) to track the
current tab.