Connect to normal "switch-page" signal.
After some consideration, it seems that connecting normally to the
"switch-page" signal from the notebook will be better. The issue with
connecting with connect_after is that when the signal handler is called,
there is no way to know the previous current page (without extra state
somewhere). Connecting normally gives the outgoing and incoming page
numbers (current is unchanged, new page is given as a parameter), so it
will be possible to save tab state, if necessary.

The only change in the "switch-page" handler is that the URI and title
must be updated from the incoming tab, not the current tab, which has
not changed at that point.