Save and restore user input for each tab.
When the current tab is switched out, any user-inputed text in the
toolbar entry is saved inside the tab object. The incoming tab is also
checked for any saved text that was stored when it was switched out.

There were some changes to how the URI is updated. It is now possible to
provide an "overwrite" flag to update_uri* that will forcefully
overwrite the toolbar entry's user-inputed text. Of course, this text
should already be saved somewhere. This is needed for when the outgoing
tab had user inputed text and the incoming does not, so the URI is set
from the tab's web-view.

Also, when the URI is set from saved user inputed text, the
entry_modified flag needs to be set in the toolbar. So there was a
"user" boolean flag added to the set_entry_uri function to control the
modified flag.