Enable closing of tabs by clicking the 'x' button.
The tab-label's close button is grabbed inside the notebook when a new
page is created and its click event is propagated (standard procedure).
In order to disconnect the tab-label click event handler, the
GtkContainer's remove virtual function is overriden because the
GtkNotebook virtual function is called after the tab is removed (so
there is no easy way to get the tab-label).

In the window's on_tab_removed there was some discussion of adding an
undo-close-tab feature, similar to most browsers. This was motivated by
the saving and storing of tab state in on_tab_changed. When a tab is
closed, it could be saved in an undo list. This undo list could be a
list of uri, user input (text), back-forward history, and zoom & scroll
settings. Actually, this is a whole other can of worms, since it seems
like a BrowserTabState (or WebkitWebViewState?) object is what we really
want.