Update toolbar buttons to match back-forward list.
The toolbar back and forward buttons are updated whenver the current
tab's web-view has a change to its back-forward list. The buttons are
also updated when the tab is changed.

It may be easier in the future to encapsulate some of the useful tab
state in a TabState object (not an enum?) that can be grabbed whenever
required and used by the window to update all the widgets.