Implement tab state and show loading spinner.
Currently the tab can have two states: loading, or normal. In the
loading state, we show the progress spinner in the tab-label.

The tab state is implemented as an enum that is stored in the object
instance. However, it is fudged as a property with the get_property (it
can't be set though). Maybe it would be better to make it genuine
property, or have other widgets attach to the web-view's load-changed
signal? Or maybe not.