Allow saving of persistent cookies.
Cookies are saved in a local SQLite file in the Mozilla format.

Originally, the cookie code was placed in browser_app_init(), but this
caused many errors in the program relating the a Visual object? Anyways,
the correct place to put initialization code is in the startup function
of the app. startup is only called once when the application is first
started.