Add basic favicon support in tab-labels.
In order for WebKit to load favicons for a web-view, it is necessary to
set the favicon database directory. Connecting to the "notify::favicon"
signal of the web-view only has an effect once the favicon database
directory has been set. (!)

Several TODOs related to refactoring code were added in the tab-label
class. The most important is to add unreffing of the GdkPixbuf used for
the favicon in the tab-label icon (memory leak otherwise).