Add basic action & accel support for app.
App actions are loaded when the app does its startup, and accels are
also set for each action. Actions are set on the GActionMap (app), so
when there are window actions in the future, these will need to be
loaded in the window's startup function (or equivalent). However, all
accels are set on the app and this is the reason why when setting accels
the full detailed action name is used ("app.quit" vs. "quit").