Add fullscreen action to window.
The window can now be fullscreen-ed and unfullscreen-ed with the
"win.fullscreen" action.

In the fullscreen state-change handler, win_toggle_fullscreen, the
fullscreen state is set at the end of the function in all cases. It
would be more robust to connect to the "window-state-event" signal for
the window and set the action state in the handler. This is because it
is not guaranteed that the fullscreen request will be granted by the
window manager. If the current implementation is kept, it could be
possible for the action state and window state to be inconsistent.