Remove propagation signals and use actions instead
Butons in the toolbar and notebook now attach themselves to an action on
the window or app. This could be done in the code with
gtk_actionable_set_action_name(...), but it is even easier to set the
action_name property of the GtkButtons in the UI XML files. The prefix
("app" or "win") is required in the UI files to find the correct action
group.